--- carrot/src/main/kotlin/waffle/team6/global/auth/JwtAuthenticationTokenFilter.kt
@@ -1,21 +1,0 @@
-package waffle.team6.global.auth
-
-import org.springframework.stereotype.Component
-import org.springframework.web.filter.OncePerRequestFilter
-import javax.servlet.FilterChain
-import javax.servlet.http.HttpServletRequest
-import javax.servlet.http.HttpServletResponse
-
-@Component
-class JwtAuthenticationTokenFilter(
-
-):OncePerRequestFilter() {
-    override fun doFilterInternal(
-        request: HttpServletRequest,
-        response: HttpServletResponse,
-        filterChain: FilterChain
-    ) {
-        TODO("Not yet implemented")
-    }
-
-}
