+++ carrot/src/main/kotlin/waffle/team6/global/auth/SecurityConfig.kt
@@ -1,15 +1,19 @@
 package waffle.team6.global.auth
 
 import org.springframework.context.annotation.Bean
+import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter
 import org.springframework.security.config.http.SessionCreationPolicy
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
 import org.springframework.security.crypto.password.PasswordEncoder
 
+@EnableGlobalMethodSecurity(prePostEnabled = true)
+@EnableWebSecurity
 class SecurityConfig(
     private val jwtAuthenticationEntryPoint: JwtAuthenticationEntryPoint,
-    private val jwtAuthenticationTokenFilter: JwtAuthenticationTokenFilter,
+    private val jwtAuthenticationFilter: JwtAuthenticationFilter,
 ): WebSecurityConfigurerAdapter() {
 
     @Bean
@@ -25,5 +29,6 @@ class SecurityConfig(
             .and()
             .authorizeRequests()
             .antMatchers("/api/v1/users/signin").permitAll()
+        //TODO add filter
     }
 }
