+++ carrot/src/main/resources/application.yml
@@ -28,6 +28,11 @@ spring:
       max-file-size: 100MB
       max-request-size: 100MB
 
+
+jwt:
+  header: Authorization
+  jwt-secret-key: wafflestudio
+  jwt-expiration-in-ms: 10800000
 ---
 
 spring:
